SensioLabs CS fixes
