+++ Behat/Context/Page/AdminPanel.php
@@ -9,6 +9,8 @@
 
 namespace FSi\Bundle\AdminBundle\Behat\Context\Page;
 
+use Behat\Mink\Element\NodeElement;
+
 class AdminPanel extends Page
 {
     protected $path = '/admin/';
@@ -52,7 +54,7 @@ class AdminPanel extends Page
 
         $linkNodes = $this->findAll('css', 'li#language > ul > li');
 
-        return array_filter(array_map(function($element) {
+        return array_filter(array_map(function(NodeElement $element) {
             return $element->getText();
         }, $linkNodes));
     }
