+++ Exception/ExceptionInterface.php
