+++ Factory/Worker/DoctrineWorker.php
@@ -18,7 +18,7 @@ class DoctrineWorker implements Worker
     /**
      * @param ManagerRegistry $managerRegistry
      */
-    function __construct(ManagerRegistry $managerRegistry)
+    public function __construct(ManagerRegistry $managerRegistry)
     {
         $this->managerRegistry = $managerRegistry;
     }
