+++ vagrant/.gitignore
